+++ src/main/java/com/mbc/clickclinic/dao/MedicamentRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MedicamentRepository extends JpaRepository<Medicament,Long> {
+public interface MedicamentRepository extends JpaRepository<Medicament,Integer> {
 
     public Medicament findMedicamentByLibelle(String libelle);
     public Medicament findMedicamentByCodeATC(String codeATC);
