+++ src/main/java/com/mbc/clickclinic/dao/OrdonnanceRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrdonnanceRepository extends JpaRepository<Ordonnance,Long> {
+public interface OrdonnanceRepository extends JpaRepository<Ordonnance,Integer> {
 }
