+++ src/main/java/com/mbc/clickclinic/dao/SalleDattentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SalleDattentRepository extends JpaRepository<SalleDattente,Long> {
+public interface SalleDattentRepository extends JpaRepository<SalleDattente,Integer> {
 }
