+++ src/main/java/com/mbc/clickclinic/dao/SecretaireRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SecretaireRepository extends JpaRepository<Secretaire,Long> {
+public interface SecretaireRepository extends JpaRepository<Secretaire,Integer> {
 
     public Secretaire findSecretaireByNom(String nom);
 }
