+++ src/main/java/com/mbc/clickclinic/service/CongesService.java
@@ -8,6 +8,6 @@ import java.util.List;
 public interface CongesService {
 
     public Conges createConges(Conges conges);
-    public Conges getCongesById(Long id);
+    public Conges getCongesById(int id);
     public List<Conges> getCongesByMedecin(Medecin medecin);
 }
